save_data/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery
google-cloud-logging
